requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
dj-email-url==1.0.5
Django==3.2
django-cache-url==3.4.0
django-debug-toolbar==3.2.1
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
environs==9.3.2
marshmallow==3.15.0
packaging==21.3
phonenumbers==8.12.46
Pillow==8.2.0
pyparsing==3.0.7
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
